bugfix: S3C-2541 remove existence pre-check before Add
It is more correct to allow duplicates to be inserted than silently
dropping them, because deleting the duplicates could otherwise kick
out random keys, breaking the no-false-negative invariant. The
downside is that the filter can only accept a limited number of
duplicate keys, but for the Metadata use-case there is no duplicate.